requirements.txt
aws-python-helper==0.34.0
beautifulsoup4>=4.14.3
boto3>=1.39.0
google-genai>=1.68.0
pydantic>=2.12.5
pydantic-settings>=2.13.1
pymongo==4.6.1
pytest>=9.0.0
pytest-cov>=6.2.0
python-dotenv>=1.2.2
requests>=2.33.0
urllib3>=2.5.0
